move code to check country out of main
